% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging_function.R
\name{aging_fun}
\alias{aging_fun}
\title{Compute the Change in Population Sizes due to Aging and Births}
\usage{
aging_fun(age.cat)
}
\arguments{
\item{age.cat}{age.cat is a vector of length 2 defined in load_start that
defines the widths of the two age categories (currently 10 and 15 y)}
}
\description{
Output from this function is stored in the `gc_env` environment.
Outputs include the `aging`, `p.nsa`, `births`, `births.sa`,
and `births.nsa` matrices.
}
