% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier_funs.R
\name{bezier}
\alias{bezier}
\title{Import the Bezier function from the bezier package, and not Hmisc}
\usage{
bezier(t, p, start = NULL, end = NULL, deg = NULL)
}
\description{
Import the Bezier function from the bezier package, and not Hmisc
}
