% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_calibration_outcomes.R
\name{burn_and_thin_traces_from_path}
\alias{burn_and_thin_traces_from_path}
\title{Burn and Thin the Traces from a Directory Containing Calibration Data Files}
\usage{
burn_and_thin_traces_from_path(path, burn, thin)
}
\description{
Burn and Thin the Traces from a Directory Containing Calibration Data Files
}
