% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_gcRegional}
\alias{calibrate_gcRegional}
\title{Run Calibration, Render Prior/Posterior Plots, and Save Outputs}
\usage{
calibrate_gcRegional(site, niterations, output_directory, calibration_filename,
  plot_filename, sd.factor = 100)
}
\description{
Run Calibration, Render Prior/Posterior Plots, and Save Outputs
}
