% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_analysis.R
\name{construct_key_populations}
\alias{construct_key_populations}
\title{Consruct the List of Key Populations which will be used in the intervention analysis.}
\usage{
construct_key_populations()
}
\description{
This function is called in load_start so that the key_populations will be available
for the intervention_analysis.
}
