% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pred_function.R
\name{create_gcSim_param_env}
\alias{create_gcSim_param_env}
\title{Create Parameters for Calling the gcSim cppFunction}
\usage{
create_gcSim_param_env(theta, interv = "base_case")
}
\description{
Create Parameters for Calling the gcSim cppFunction
}
