% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_calibration_outcomes.R
\name{extract_traces}
\alias{extract_traces}
\title{Extract Traces from Calibrations}
\usage{
extract_traces(calibrations)
}
\arguments{
\item{calibrations}{A list of calibration outputs from my_mcmc}
}
\description{
Extract Traces from Calibrations
}
