% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_environment.R
\name{gc_assign}
\alias{gc_assign}
\title{Easier Named Assignment into the gc_env Environment}
\usage{
gc_assign(var)
}
\description{
The end result of this function is no different than
using `gc_env$var <- var` directly. Its utility is that it eliminates
the need to type the variables' names twice.
}
