% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_environment.R
\docType{data}
\name{gc_env}
\alias{gc_env}
\title{The gcRegional Variable Environment}
\format{An object of class \code{environment} of length 2.}
\usage{
gc_env
}
\description{
This is the environment where variables and parameters necessary for the
gcRegional model will be stored. In prior versions of this model, many
variables were stored in the global scope -- any instances of those variables'
assignment and retrieval to the global scope have been replaced with assignment
and retrieval from gc_env.
}
\section{initial_files}{
 The following gc_env objects are assigned in the
  initial_files function.

\itemize{
\item s.dist.dat.m: Partnership distribution by subpop, Males, NSFG, not currently using for fitting
\item s.dist.dat.f: Partnership dist'n by subpop, Females, NSFG, not currently using for fitting
\item age.dist.dat: Proportion of partnerships with person of same age group, NSFG (national level estimates)
\item prev.extra.cat.dat: Prevalence, NHANES, not using for fitting but keep for comparison
\item diag.rr: RR of diagnosis rate for black and Hispanic, relative to overall rate for given age and sex
\item diag.age.sex.rate: Case rate data by age sex only, 2002-2016, sd calculated assuming +/-20%
\item p.symp.dat:
\item p.msm.dat: Proportion of male diagnosed cases in SSuN occuring in MSM, by age group
\item diag.subpop.rate: Rates by subpop for Baltimore, 2011-2016
\item priors: Prior distributions (parameters for probablility distributions for each input parameter)
\item theta: Vector of model parameter values
\item dat.s.dist: Proportion of within group mixing for each subpopulation, from NSFG data (not currently using for calibration)
\item var.s.dist.dat: Variance for mixing
\item s.dist.sd:
\item rr.diag.subpop: mean values of rr reported case by subpop
\item rr.diag.subpop.sd: sd for rr reported case by subpop (assuming +/- 20% and normal dist)
\item p.symp.ssun: Proportion of treated cases that are symptomatic
\item var.symp.ssun: Variance estimated from fitting data to beta distribution
\item p.msm.ssun: Proportion of males cases diagnosed in MSM, 2010-2016 SSuN
\item var.p.msm.ssun: Variance assuming beta distn and 95% CI=+/- 20% of mean
\item diag.rate: diagnosed case rates by age and sex only
\item diag.rate.sd: SD for diagnosed case data
\item prev.nhanes: GC prevalence 1999-2008 at national level, not currently using for model fitting
\item var.prev.nhanes:
\item sd.prior: starting value for standard deviation associated with each parameter, adapted during fitting
\item prior.param1: first parameter describing probablity distributions
\item prior.param2: second parameter describing probability distributions
}
}

\section{pop_calc}{
 The following gc_env objects are assigned in the
  pop_calc function.

\itemize{
\item p.k: Distribution of population by low and high activity status
\item p.sa.array: Proportion of sexually active
\item n.dist: Population size
\item n.dist.sa: Population size for sexually active
\item p.dist: Population distribution within subpopulation and age group
\item p.s.dist: Relative sizes of different subpopulations - assume that mixing outside of subpopulation is proportionate to number of individuals in each subpopulation, excluding MSM
\item n.s.dist: Population sizes by sex (j) and subpopulation (i)
\item n.s.dist.sa: Population sizes by sex (j) and subpopulation (i) for sexually active population
\item n.s.pop: Population sizes, minus MSM
\item n.s.pop.sa: Population sizes, minus MSM, for sexually active population
\item n.i: (Population size by sex, subpopulation, AC, age) M, subpop=1:  age=AC=1, age=1 AC=2, age=2 AC=2; M subpop=2: age=AC=1,etc. Then F, subpop1...
\item n.sa: Size of sexually active population
\item n.nsa: Size of not sexually active population
}
}

\keyword{datasets}
