% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_environment.R
\name{gc_get}
\alias{gc_get}
\title{Easier Variable Retrieval from the gc_env Environment}
\usage{
gc_get(varname, env = parent.env(environment()))
}
\description{
The functionality of gc_get is no different than performing
`varname <- gc_env$varname`. This function keeps the user
from needing to type the varname twice.
}
