% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcRegional.R
\docType{package}
\name{gcRegional-package}
\alias{gcRegional}
\alias{gcRegional-package}
\title{A Regional Model of Gonorrhea in Baltimore and San Francisco}
\description{
A Regional Model of Gonorrhea in Baltimore and San Francisco
}
\author{
\strong{Maintainer}: Christian Testa \email{ctesta@hsph.harvard.edu}

Authors:
\itemize{
  \item Ashleigh Tuite \email{atuite@hsph.harvard.edu}
}

}
