% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pred_function.R
\name{get_simulation_statistics}
\alias{get_simulation_statistics}
\title{Render Statistics from a Simulation}
\usage{
get_simulation_statistics(e)
}
\description{
Render Statistics from a Simulation
}
