% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_plotting.R
\name{infection_str}
\alias{infection_str}
\title{This function converts y/z/all -> "Symptomatic ", "Asymptomatic ", ""
for the purpose of titling relative burden plots.}
\usage{
infection_str(which_infections)
}
\description{
This function converts y/z/all -> "Symptomatic ", "Asymptomatic ", ""
for the purpose of titling relative burden plots.
}
