% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_files.R
\name{initial_files}
\alias{initial_files}
\title{Load Data from Initial Files}
\usage{
initial_files()
}
\description{
This reads the data in the package's inst/extdata folder and
places that data into the gc_env environment.
}
