% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_analysis.R
\name{intervention_analysis_key_metrics}
\alias{intervention_analysis_key_metrics}
\title{Construct Key Metrics from a Simulation out of gcSim}
\usage{
intervention_analysis_key_metrics(sol)
}
\description{
Let's take the simulation data from gcSim and save only some key metrics
in a data frame.
}
