% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes_2.r
\name{layout_target_and_fit_plots}
\alias{layout_target_and_fit_plots}
\title{Layout the Target and Fit Plots}
\usage{
layout_target_and_fit_plots(l)
}
\description{
Later I may turn this into a function that renders PDFs using these layouts.
Or I might return a list of page-by-page layouts for rendering in a PDF.
}
