% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes_2.r
\name{list_target_and_fit_plots}
\alias{list_target_and_fit_plots}
\title{Plot Calibration Targets}
\usage{
list_target_and_fit_plots()
}
\description{
Construct a named list of ggplot objects, each of which depicts one of the
pairs of data targets and the model fit against it.
}
