% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{load_BA16_data}
\alias{load_BA16_data}
\title{Read in the SSuN 2016 Data Set from Baltimore}
\usage{
load_BA16_data()
}
\description{
This function is called immediately after its construction
so that the BA16_data is available for the gcRegional package.
}
