% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reading.R
\name{load_BA17_data}
\alias{load_BA17_data}
\title{Read in the SSuN 2017 Data Set from Baltimore}
\usage{
load_BA17_data()
}
\description{
This function is called immediately after its construction
so that the BA16_data is available for the gcRegional package.
}
