% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_calibration_outcomes.R
\name{load_calibration_rds_files}
\alias{load_calibration_rds_files}
\title{Load Calibration RDS files}
\usage{
load_calibration_rds_files(path)
}
\arguments{
\item{path}{A directory containing RDS files which are the output of my_mcmc}
}
\description{
Given a path, this function reads in each calibration file
and returns a list containing the data from each RDS
file contained in that path.
}
