% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_start_conditions.R
\name{load_start}
\alias{load_start}
\title{Load Starting Conditions}
\usage{
load_start(site = NA)
}
\description{
When running load_start() the user will be prompted to specify the site location.
After doing so, initial_files() will run in order to load the initial parameters
and variables that correspond to the specified site. The site and variables
loaded will be stored in the gc_env environment.
}
