% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMCMC.R
\name{myMCMC}
\alias{myMCMC}
\title{Metropolis Hastings MCMC}
\usage{
myMCMC(target, init.theta, proposal.sd = NULL, n.iterations, covmat = NULL,
  limits = list(lower = NULL, upper = NULL), adapt.size.start = NULL,
  adapt.size.cooling = 0.99, adapt.shape.start = NULL,
  print.info.every = n.iterations/100, verbose = FALSE,
  max.scaling.sd = 50, acceptance.rate.weight = NULL,
  acceptance.window = NULL)
}
\description{
This is a slight adaptation of mcmcMH from fitR package
 prevents 'covariance matrix not positive definite' error
 see fitR package for documentation.
}
\details{
Original Function located here:
 https://sbfnk.github.io/fitR/reference/mcmcMH.html

References for Adaptive MCMC:

https://people.eecs.berkeley.edu/~jordan/sail/readings/andrieu-thoms.pdf
http://probability.ca/jeff/ftpdir/adaptex.pdf
https://projecteuclid.org/download/pdf_1/euclid.bj/1080222083
}
