% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_likelihood.R
\name{plot_likelihood_components}
\alias{plot_likelihood_components}
\title{Plot Uniform Likelihood Region}
\usage{
plot_likelihood_components()
}
\value{
A list of ggplots which represent the likelihood
region around each data point.
}
\description{
Plot Uniform Likelihood Region
}
