% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outputs.R
\name{plot_posteriors}
\alias{plot_posteriors}
\title{Plot Calibration Targets and Priors/Posteriors}
\usage{
plot_posteriors(output_directory, filename = paste0("Calibration_Plots_",
  gc_env$site, ".pdf"))
}
