% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_plotting.R
\name{popmatrix}
\alias{popmatrix}
\title{Population Indicator Matrix}
\description{
This matrix's column names contain the different demographic categories,
i.e. race-ethnicity, gender, activity rate, and age category.
Its rows' indices refer to compartments in the model.
For example, one can query the matrix for what demographics
a compartment represents by looking at popmatrix[i, ], or
similarly one can look at which compartments represent a
particular demographic by looking at popmatrix[, i].
}
\details{
Be careful not to use the Female MSM! They are in the model
for consistency, so that we can loop through populations
using a selection format like (i*j*k*l), but they are
always at 0 population size!
}
