% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_weighted_likelihood.R
\name{population_weighted_likelihood}
\alias{population_weighted_likelihood}
\title{Population Weighted Likelihood Function}
\usage{
population_weighted_likelihood(pred)
}
\description{
Population Weighted Likelihood Function
}
