% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_calibration_outcomes.R
\name{predictions_and_plots_from_files}
\alias{predictions_and_plots_from_files}
\title{Render Predictions and Plots from Calibration Files}
\usage{
predictions_and_plots_from_files(ba_path, sf_path, output_path, burn, thin,
  sample_n)
}
\description{
Render Predictions and Plots from Calibration Files
}
