% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_outcomes_2.r
\name{prepare_data_for_prior_posterior_target_and_fit_plots}
\alias{prepare_data_for_prior_posterior_target_and_fit_plots}
\title{Prepare Data for Prior, Posterior, Target, and Model Fit Plots}
\usage{
prepare_data_for_prior_posterior_target_and_fit_plots()
}
\description{
Prepare Data for Prior, Posterior, Target, and Model Fit Plots
}
