% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_likelihood.R
\name{prepare_for_plotting_likelihoods}
\alias{prepare_for_plotting_likelihoods}
\title{Prepare plot_environment with necessary computations and values}
\usage{
prepare_for_plotting_likelihoods()
}
\description{
Prepare plot_environment with necessary computations and values
}
