% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_posterior_means.R
\name{remove_national_parameters}
\alias{remove_national_parameters}
\title{Remove Parameters which are Fixed for use with National Posterior Mean Values}
\usage{
remove_national_parameters(theta)
}
\description{
Remove Parameters which are Fixed for use with National Posterior Mean Values
}
