% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_posterior_means.R
\name{replace_national_parameter_medians}
\alias{replace_national_parameter_medians}
\title{Replace National Posterior Mean Values}
\usage{
replace_national_parameter_medians(theta)
}
\description{
Replace National Posterior Mean Values
}
