% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_likelihood.R
\name{save_plots_from_plot_environment}
\alias{save_plots_from_plot_environment}
\title{Save Plots from plot_environment}
\usage{
save_plots_from_plot_environment()
}
\description{
Save Plots from plot_environment
}
