% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_plotting.R
\name{select_populations}
\alias{select_populations}
\title{Select Population Names by their Indices}
\usage{
select_populations(ijkl_subset = c("i", "j", "k", "l"))
}
\arguments{
\item{ijkl_subset}{Choose a subset of c('i', 'j', 'k', 'l') to pick subpopulations split by
the characteristics corresponding to each of these letters. For the correspondence between
these letters and the characteristics, see the gc_env$popnames object.}
}
\description{
This function converts a subset of c('i', 'j', 'k', 'l')
into a list of all possible intersections between each
of the designated populations.
}
