% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions_analysis.R
\name{simulate_interventions}
\alias{simulate_interventions}
\title{Simulate and Interventions for a Trace}
\usage{
simulate_interventions(interventions)
}
\description{
Simulate and Interventions for a Trace
}
