% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_plotting.R
\name{tidy_prevalence_by_demographic}
\alias{tidy_prevalence_by_demographic}
\title{Calculate the Percentage of Burden for Subpopulations}
\usage{
tidy_prevalence_by_demographic(sol, ijkl_subset = c("i", "j", "k", "l"),
  which_infections = "all", relative = TRUE, years = 2010:2016)
}
\arguments{
\item{sol}{The output of the C++ gcSim function, stored by prediction_epi as gc_env$sol}

\item{ijkl_subset}{Choose a subset of c('i', 'j', 'k', 'l') to pick subpopulations split by
the characteristics corresponding to each of these letters. For the correspondence between
these letters and the characteristics, see the gc_env$popnames object.}

\item{which_infections}{Specify "all" to tabulate both symptomatic and asymptomatic infections,
"y" for symptomatic, and "z" for symptomatic only.}

\item{relative}{If relative is TRUE, then each subpopulation is divided by the total number of
infections so that the output values represent each subpopulations percentage of the total
infections. If relative is FALSE, raw numbers of infections are output.}

\item{years}{Specify the vector of years to output prevalence data from.
The usual options for the years parameter are
2010:2016  for the data period,
2002:2016  for the calibration period, or
1942:2016  for the entire simulation period.}
}
\description{
We start with sol, the output of gcSim which is stored in gc_env by the
gcRegional::prediction_epi. sol is a dataframe which contains the
population sizes of each compartment at each timestep. For this function, we
are interested only in infections, and we subset sol for the
symptomatic/asymptomatic/all infections based on the which_infections
parameter. Using the ijkl_subset this function uses the
select_populations to group subpopulations and calculate the number
ofinfections experienced by a given subpopulation. Then the number of
infections of each subpopulation is divided by the total number of infections
to calculate the relative burden of each subpopulation.
}
