% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_plotting.R
\name{tidy_prevalence_dfs_for_many_simulations}
\alias{tidy_prevalence_dfs_for_many_simulations}
\title{tidy_prevalence_dfs_for_many_simulations applies tidy_prevalence_by_demographic to many
simulations}
\usage{
tidy_prevalence_dfs_for_many_simulations(trace_sample, ijkl_subset = c("i",
  "j", "k", "l"), relative = T, years = 2010:2016,
  which_infections = "all", interv = "base_case")
}
\description{
@inheritParams tidy_prevalence_by_demographic
 @param sample_size The number of simulations to use for generating tidy prevalence data.
 interv An intervention plan that changes how screening is implemented after 2016.
 @param trace_sample A dataframe with rows of theta-values sampled from the trace of
 an MCMC calibration of the gcRegional model.
}
\details{
This function outputs a list of dataframes, each of which is generated individually
 by tidy_prevalence_by_demographic.
}
