% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_likelihood.R
\name{uniform_likelihood}
\alias{uniform_likelihood}
\title{Construct a Uniform Likelihood Function}
\usage{
uniform_likelihood(pred, N = 5)
}
\value{
a value 0 or 1
}
\description{
In order to come up with trust-able results, we want to
try creating a version of the likelihood function
which is either 1 in a range around each data point but
0 everywhere outside each range.
}
\details{
First, a list of the likelihood components:
c(
  'Age Assortative Mixing',
  'Proportion of Male Cases in MSM 15-24 y',
  'Proportion of Male Cases in MSM 25-39 y',
  'Proportion of Diagnosed Cases that are Symptomatic',
  'Reported Cases',
  'Reported Cases Relative Risk',

)
}
